requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
dnspython==1.16.0
et-xmlfile==1.0.1
jdcal==1.4.1
more-itertools==8.3.0
mysql-connector-python==8.0.19
openpyxl==3.0.3
packaging==20.4
pluggy==0.13.1
protobuf==3.6.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.9.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.4
